requirements.txt
alembic==1.12.0
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.103.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.3.0
pydantic_core==2.6.3
pytest==7.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.23.2
